+++ springproject/src/main/java/it/academy/model/Unit.java
@@ -14,7 +14,7 @@ public class Unit {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long unitId;
+    private Integer unitId;
 
     private String unitName;
 }
